delete deneme2

+++ uninstall.php
@@ -1,11 +1,6 @@
 <?php 
-
-register_uninstall_hook(__FILE__, 'deleteAction');
-
-function deleteAction () {
+    if( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) exit();
     global $wpdb;
     $tableName = $wpdb->prefix . "orderalertifylogs"; 
-    $wpdb->query( "DROP TABLE IF EXISTS $tableName" );
-}
-
+    $wpdb->query( "DROP TABLE IF EXISTS $tableName");
 ?>
